LayoutTests/editing/execCommand/indent-images-3.html

<!DOCTYPE html>
<html>
<body><span contenteditable><img><img></span>
<script src="../../resources/dump-as-markup.js"></script>
<script>

if (window.testRunner)
    testRunner.dumpAsText();

while (script = document.querySelector('script'))
    script.parentNode.removeChild(script);

getSelection().setPosition(document.getElementsByTagName('span')[0], 2);
document.execCommand('indent');

Markup.description('This test indents insides a document with exactly two image elements.\n'
+ 'Indentation should succeed even if the root editable element is inline.');
Markup.dump(document.body);

</script>
</body>
</html>
