safari-600.7-branch/LayoutTests/editing/execCommand/indent-images-3-expected.txt

This test indents insides a document with exactly two image elements.
Indentation should fail because the root editable element is inline.
| <span>
|   contenteditable=""
|   <img>
|   <img>
| "
"
| "
"
